#56 - refactored mediainfo to information package
