+++ src/server/pkg/information/command_test.go
@@ -1,22 +1,19 @@
-package media
+package information
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-
 	"github.com/stretchr/testify/require"
 )
 
 func TestInfoFromFile(t *testing.T) {
-	info, err := RunMediaInfo("./test_data/sample1.m4a")
+	info, err := mediaInfoFromFile("./test_data/sample1.m4a")
 	require.NoError(t, err, "file should be present")
 	require.NotNil(t, info, "media info should return an object")
 
-	assert.EqualValues(t, "AAC", info.AudioFormatList)
 	assert.EqualValues(t, "./test_data/sample1.m4a", info.CompleteName)
 	assert.EqualValues(t, "./test_data", info.FolderName)
-	assert.EqualValues(t, "", info.FileName)
 	assert.EqualValues(t, "m4a", info.FileExtension)
 	assert.EqualValues(t, "MPEG-4", info.Format)
 	assert.EqualValues(t, "MPEG-4", info.CommercialName)
