+++ src/server/pkg/information/mediainfo_test.go
@@ -1,4 +1,4 @@
-package media
+package information
 
 import (
 	"testing"
