+++ src/server/pkg/information/test_data/mediainfo1.txt
